requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
websockets==13.1

# Pydantic
pydantic==2.10.5
pydantic-settings==2.7.1

# SAM3 dependencies (will use parent sam3 package)
# torch==2.7.0
# torchvision
# PIL

# Image/Video processing
opencv-python==4.10.0.84
Pillow==11.1.0
imageio==2.37.0
imageio-ffmpeg==0.5.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1

# Monitoring and Logging
prometheus-client==0.21.0
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1  # for testing

# Development
black==24.10.0
flake8==7.1.1
mypy==1.14.1
